move persistJsonFold
