+++ apps/studio/src/components/sidebar/JsonViewer.vue
@@ -88,7 +88,7 @@ import {
 } from "@/lib/data/jsonViewer";
 import { mapGetters } from "vuex";
 import { EditorMarker, LineGutter } from "@/lib/editor/utils";
-import { persistJsonFold } from "@/lib/editor/plugins/persistJsonFold";
+import { persistJsonFold } from "@/lib/editor/extensions/persistJsonFold";
 import { partialReadonly } from "@/lib/editor/extensions/partialReadOnly";
 import JsonViewerUpsell from '@/components/upsell/JsonViewerSidebarUpsell.vue'
 import rawLog from "@bksLogger";
