+++ apps/studio/src/lib/editor/extensions/persistJsonFold.ts
